+++ client/src/components/academics/academics.css
@@ -1,91 +1,99 @@
+/* Global Page Styling */
 .page {
     display: flex;
     flex-direction: column;
-    align-items: center; /* Center align horizontally */
+    align-items: center;
     padding: 20px;
     box-sizing: border-box;
-    background-color: #f4f7f8; /* Light background for the entire page */
-    min-height: 100vh; /* Ensure the page takes full viewport height */
+    background-color: #f4f7f8;
+    min-height: 100vh;
 }
 
+/* Dropdown Container Styling */
 .dropdown-container {
-    position: relative; /* Ensure the dropdown menu is positioned relative to this container */
-    margin-bottom: 10px; /* Spacing between dropdowns */
-    width: 100%; /* Ensure full width */
-    max-width: 400px; /* Limit the width of dropdowns */
+    position: relative;
+    margin-bottom: 10px;
+    width: 100%;
+    max-width: 400px;
 }
 
+/* Dropdown Button Styling */
 .dropdown-button {
-    text-align: center;
     width: 100%;
-    max-width: 400px; /* Limit the width */
+    max-width: 400px;
     height: 45px;
-    background-color: #e9eff1; /* Slightly different shade for better visual appeal */
-    border-radius: 25px; /* Slightly more rounded corners */
-    margin: 5px 0; /* Vertical margin */
+    background-color: #e9eff1;
+    border-radius: 25px;
+    margin: 5px 0;
     cursor: pointer;
-    border: 1px solid #ddd; /* Border for better separation */
+    border: 1px solid #ddd;
     outline: none;
-    position: relative; /* Position relative for arrow positioning */
+    text-align: center;
+    position: relative;
     transition: background-color 0.2s, transform 0.2s;
 }
 
 .dropdown-button:hover {
-    background-color: #dbe7e9; /* Slightly darker on hover */
+    background-color: #dbe7e9;
     font-weight: bold;
     transform: scale(1.05);
 }
 
+/* Dropdown Arrow Styling */
 .arrow {
     margin-left: 10px;
-    font-size: 12px; /* Adjust size if needed */
-    vertical-align: middle; /* Align vertically with text */
+    font-size: 12px;
+    vertical-align: middle;
 }
 
+/* Dropdown Menu Styling */
 .dropdown-menu {
     position: absolute;
-    top: 100%; /* Position below the button */
+    top: 100%;
     left: 0;
     width: 100%;
     background-color: #fff;
     border: 1px solid #ddd;
     border-radius: 5px;
     box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
-    z-index: 1000; /* Ensure menu is on top */
+    z-index: 1000;
     max-height: 250px;
-    overflow-y: auto; /* Scroll if too many items */
+    overflow-y: auto;
     padding: 0;
     margin: 0;
 }
 
+/* Dropdown Item Styling */
 .dropdown-item {
     width: 100%;
-    padding: 12px 15px; /* Slightly more padding for better touch target */
+    padding: 12px 15px;
     background-color: #fff;
     border: none;
     text-align: left;
     cursor: pointer;
     transition: background-color 0.2s;
-    font-size: 14px; /* Adjust font size if needed */
+    font-size: 14px;
 }
 
 .dropdown-item:hover {
-    background-color: #f2f4f7; /* Slightly lighter on hover */
+    background-color: #f2f4f7;
 }
 
 .dropdown-item.active {
-    background-color: #cce5ff; /* Lighter blue for active item */
+    background-color: #cce5ff;
     font-weight: bold;
 }
 
+/* Non-Toggle Button Container */
 .non-toggles {
     display: flex;
     flex-direction: column;
     align-items: center;
-    width: 100%; /* Ensure it stretches to the container width */
-    max-width: 400px; /* Limit the width */
+    width: 100%;
+    max-width: 400px;
 }
 
+/* Button Styling */
 .button {
     text-align: center;
     width: 200px;
@@ -93,15 +101,19 @@
     background-color: #f4f5fa;
     border-radius: 30px;
     padding: 2px 8px;
-    margin-top: 7px;   
+    margin-top: 7px;
     margin-left: 7px;
+    transition: 0.2s;
 }
 
-.button:hover,
 .button:hover {
     font-weight: bold;
     cursor: pointer;
     background-color: #7dafccd0;
     transform: scale(1.05);
-    transition: 0.2s;
 }
+
+h4 {
+    text-align: center;
+}
+
