--- client/src/components/academics/quicklinks.css
@@ -1,29 +1,0 @@
-.section {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.buttons {
-    text-align: center;
-    width: 200px;
-    height: 45px;
-    background-color: #f4f5fa;
-    border-radius: 30px;
-    padding: 2px 8px;
-    margin-top: 7px;
-    margin-left: 7px;
-    /* Added styles for text size and centering */
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 16px; /* Adjust the font size as needed */
-}
-
-.buttons:hover {
-    font-weight: bold;
-    cursor: pointer;
-    background-color: #7dafccd0;
-    transform: scale(1.05);
-    transition: 0.2s;
-}
